requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
mlflow<3
ipykernel
streamlit
setuptools
pytest
python-dotenv
datasets
dagshub
transformers
lime
tensorflow
tf-keras
torch